Consistently cast ctype macro arguments to int via unsigned char

+++ src/KRB4_fmt_plug.c
@@ -149,9 +149,9 @@ static int hex_decode(char *src, unsigned char *dst, int outsize)
 	pe = src + strlen(src);
 	qe = dst + outsize;
 
-	for (p = src, q = dst; p < pe && q < qe && isxdigit(((unsigned char)(*p))); p += 2) {
-		ch = tolower(((unsigned char)(p[0])));
-		cl = tolower(((unsigned char)(p[1])));
+	for (p = src, q = dst; p < pe && q < qe && isxdigit((int)(unsigned char)*p); p += 2) {
+		ch = tolower((int)(unsigned char)p[0]);
+		cl = tolower((int)(unsigned char)p[1]);
 
 		if ((ch >= '0') && (ch <= '9')) ch -= '0';
 		else if ((ch >= 'a') && (ch <= 'f')) ch -= 'a' - 10;
